requirements.txt
attrs==25.4.0
Cartopy==0.25.0
certifi==2025.11.12
cffi==2.0.0
cfgrib==0.9.15.1
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
eccodes==2.44.0
eccodeslib==2.44.0.5
eckitlib==1.32.2.5
fckitlib==0.14.0.5
findlibs==0.1.2
flexcache==0.3
flexparser==0.4
fonttools==4.60.1
herbie-data==2025.11.2
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
MetPy==1.7.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
Pint==0.25.2
platformdirs==4.5.0
pooch==1.8.2
pycparser==2.23
pyparsing==3.2.5
pyproj==3.7.2
pyshp==3.0.2.post1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
shapely==2.1.2
six==1.17.0
threadpoolctl==3.6.0
toml==0.10.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xarray==2025.11.0
